Move StaticArrayInterface support into a package extension
This makes the package much more lightweight, at the cost of a few extra lines
of code.
