+++ Project.toml
@@ -1,18 +1,23 @@
 name = "EllipsisNotation"
 uuid = "da5c29d0-fa7d-589e-88eb-ea29b0a81949"
-version = "1.9.0"
+version = "1.10.0"
 authors = ["Chris Rackauckas <accounts@chrisrackauckas.com>"]
 
 [deps]
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
+
+[weakdeps]
 StaticArrayInterface = "0d7ed370-da01-4f52-bd93-41d350b8b718"
+
+[extensions]
+EllipsisNotationStaticArrayInterfaceExt = "StaticArrayInterface"
 
 [compat]
 AllocCheck = "0.2"
 PrecompileTools = "1.0.1"
 SafeTestsets = "0.1"
 SciMLTesting = "1"
-StaticArrayInterface = "1.2.1"
+StaticArrayInterface = "1.4.1"
 Test = "1"
 julia = "1.10"
 
@@ -20,7 +25,8 @@ julia = "1.10"
 AllocCheck = "9b6a8646-10ed-4001-bbdc-1d2f46dfbb1a"
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 SciMLTesting = "09d9d899-5365-40a9-917a-5f67fddea283"
+StaticArrayInterface = "0d7ed370-da01-4f52-bd93-41d350b8b718"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["AllocCheck", "Test", "SafeTestsets", "SciMLTesting"]
+test = ["AllocCheck", "Test", "SafeTestsets", "SciMLTesting", "StaticArrayInterface"]
